requirements.txt
cohere
fastapi
uvicorn
python-dotenv
pydantic